+++ library/src/components-private/subscription/BuyerStudioSubscription.vue
@@ -19,7 +19,7 @@
           .display-column.full-width
             .nio-h4.text-primary-darker Max cost per 1000 records
             .nio-p.text-primary-dark {{ getCPM(subscription.details.pricing.micro_cents_usd) }}
-    .display-row.display-table.included-filters(v-if="subscription.status === 'active' || subscription.status === 'kickoff' || subscription.status === 'pending'" )
+    .display-row.display-table.included-filters
       .display-column.full-width
         .nio-h4.text-primary-darker(style="margin-bottom: 8px") Included Filters
         .applied-filters(v-if="appliedFilters && appliedFilters.length > 0")
@@ -34,7 +34,7 @@
                 display-only
               )
         .nio-p.text-primary-dark.empty(v-else) No filters applied
-    .display-row.display-table(v-if="subscription.status === 'active' || subscription.status === 'kickoff' || subscription.status === 'pending'")
+    .display-row.display-table
       .display-column.full-width
         .nio-h4.text-primary-darker(style="margin-bottom: 8px") Deliverable attributes
         .attributes