+++ library/src/components/icon/Icon.vue
@@ -67,6 +67,7 @@ import { faArchive } from './fa-icons/archive'
 import { faEllipsisV } from './fa-icons/ellipsis-vertical'
 import { faLocationArrow } from './fa-icons/location-arrow'
 import { faDownload } from './fa-icons/download'
+import { faPercent } from './fa-icons/percent'
  
 export default {
   name: 'nio-icon',
@@ -133,6 +134,7 @@ export default {
     library.add(faEllipsisV)
     library.add(faLocationArrow)
     library.add(faDownload)
+    library.add(faPercent)
     this.icon = this.iconLibrary[this.name]
     this.$emit('mounted')
   },