+++ library/src/modules/app/app-module.js
@@ -20,12 +20,16 @@ export default {
 		})
 	},	
   mounted() {
-    
+
   },
   methods: {
     nioInitializeApplication: (app) => {
       parent.postMessage({
         name: 'fetchCustomParameters',
+        payload: null
+      },"*")
+      parent.postMessage({
+        name: 'fetchAppInstallationId',
         payload: null
       },"*")
       if (window.addEventListener) {
@@ -71,7 +75,7 @@ export default {
           this.$store.dispatch('nioServices/SET_CUSTOM_APP_PARAMETERS', evt.data.payload)
           break;
         case 'appInstallationIdFetched':
-          this.$store.dispatch('nioService/SET_APP_INSTALLATION_ID', evt.data.payload)
+          this.$store.dispatch('nioServices/SET_APP_INSTALLATION_ID', evt.data.payload)
           break;
         default:
           break;