+++ CHANGELOG.md
@@ -158,3 +158,4 @@ Added **utility-download** icon
 v0.5.0
 Added optional **summary*** prop to **NioFilterGroup** and descendant components to enable non-interactable display of filter settings for listing pages
 Added **utility-percent** icon
+Added **nioAppInstallationId** to app store and associated event handler to fetch from DSM